ref(cart): separate logged-in and not logged-in `CartItem`
