+++ src/app/sn/carts/Cart/NotLoggedInUserCart/NotLoggedInUserCartItem/style.module.scss
@@ -1,0 +1,10 @@
+@use '@/styles/theme.scss' as *;
+
+.cart-item {
+  padding-bottom: 20px;
+  border-bottom: 1px solid $grey300;
+}
+
+.ingredient {
+  font-size: 1.2rem;
+}
