+++ src/app/sn/carts/Cart/QuantityInput.tsx
@@ -1,5 +1,6 @@
-import Icon from '@/components/Icon';
 import style from './style.module.scss';
+
+import Icon from '@/components/Icon';
 
 interface QuantityProps {
   quantity: number;
