+++ src/app/sn/carts/Cart/style.module.scss
@@ -16,12 +16,12 @@
 }
 
 .quantity {
-  margin-top: 10px;
+  width: 110px;
   display: flex;
   align-items: center;
   justify-content: space-between;
+  margin-top: 10px;
   padding: 7px 10px;
-  width: 110px;
   border: 3px solid $pointColor;
   border-radius: 20px;
 }
